+++ tests/e2e/specs/MainView.js
@@ -6,8 +6,7 @@ module.exports = {
     browser
       .url(process.env.VUE_DEV_SERVER_URL)
       .waitForElementVisible('#main-view', 5000)
-      .assert.elementPresent('div.row > .col-sm-4')
-      .assert.elementPresent('div.row > .col-sm-8')
+      .assert.elementPresent('div.row > .col')
       .end()
   }
 }
