+++ tests/unit/views/ContentView.spec.ts
@@ -1,48 +1,21 @@
 import { shallowMount, createLocalVue } from '@vue/test-utils'
 import ContentView from '@/views/ContentView.vue'
 import Vuex, { Store } from 'vuex'
-import SearchComponent from '@/components/search/SearchComponent.vue'
-import mutations from '@/store/mutations'
+import Vue from 'vue'
+Vue.filter('i18n', (value: string) => value) // Add dummy filter for i18n
 
 const localVue = createLocalVue()
 localVue.use(Vuex)
 
 describe('ContentView.vue', () => {
-  const filterSections = jest.fn()
-  const filterSubsections = jest.fn()
-  const loadGridVariables = jest.fn()
-  const updateSearchTerm = jest.fn()
-  const isSearchResultEmpty = jest.fn()
   const store = new Store({
-    state: { treeSelection: 3 },
-    actions: {
-      filterSections,
-      filterSubsections,
-      loadGridVariables
-    },
-    mutations: {
-      updateSearchTerm
-    },
-    getters: {
-      isSearchResultEmpty
-    }
+    state: { treeSelection: 3 }
   })
 
   it('Renders the content', () => {
     const wrapper = shallowMount(ContentView, { store, localVue })
 
     expect(wrapper.exists()).toBeTruthy()
-    expect(wrapper.find('#Content-view').exists()).toBeTruthy()
-  })
-
-  it('updates search term and dispatches filter actions when the search term changes', () => {
-    const wrapper = shallowMount(ContentView, { store, localVue })
-
-    wrapper.find(SearchComponent).vm.$emit('seachChanged', 'mini')
-
-    expect(updateSearchTerm).toHaveBeenCalled()
-    expect(filterSections).toHaveBeenCalled()
-    expect(filterSubsections).toHaveBeenCalled()
-    expect(loadGridVariables).toHaveBeenCalled()
+    expect(wrapper.find('#content-view').exists()).toBeTruthy()
   })
 })
