+++ tests/unit/views/GridView.spec.ts
@@ -69,7 +69,7 @@ describe('GridView', () => {
   describe('On creation', () => {
     it('Renders the gridView', () => {
       wrapper = shallowMount(GridView, { store })
-      expect(wrapper.find('#Grid-view').exists()).toBeTruthy()
+      expect(wrapper.find('#grid-view').exists()).toBeTruthy()
     })
   })
