+++ tests/unit/views/MainView.spec.ts
@@ -1,6 +1,7 @@
 import { shallowMount, createLocalVue } from '@vue/test-utils'
 import MainView from '@/views/MainView.vue'
 import Vuex from 'vuex'
+import SearchComponent from '@/components/search/SearchComponent.vue'
 
 describe('MainView.vue', () => {
   const localVue = createLocalVue()
@@ -9,20 +10,35 @@ describe('MainView.vue', () => {
   let actions: any
   let state: any
   const mocks: any = { '$route': { params: {} } }
+  const updateSearchTerm = jest.fn()
+  const isSearchResultEmpty = jest.fn()
+  const filterSections = jest.fn()
+  const filterSubsections = jest.fn()
+  const loadGridVariables = jest.fn()
 
   beforeEach(() => {
     state = {
+      state: { treeSelection: 3 },
       toast: { type: 'danger', message: 'i am not a message' }
     }
     actions = {
       loadVariables: jest.fn(),
       loadAssessments: jest.fn(),
+      filterSections,
+      filterSubsections,
+      loadGridVariables,
       load: jest.fn(),
       save: jest.fn()
     }
     store = new Vuex.Store({
       state,
-      actions
+      actions,
+      mutations: {
+        updateSearchTerm
+      },
+      getters: {
+        isSearchResultEmpty
+      }
     })
   })
 
@@ -41,12 +57,6 @@ describe('MainView.vue', () => {
     expect(wrapper.find('toast-component-stub').attributes('message')).toEqual('i am not a message')
   })
 
-  it('renders an save button that saves the current state', () => {
-    const wrapper = shallowMount(MainView, { store, localVue, mocks })
-    wrapper.find('#save').trigger('click')
-    expect(actions.save).toHaveBeenCalled()
-  })
-
   it('loads an order, after loading variables and assessments, if a cartId route param is present', (done) => {
     actions.loadVariables.mockReturnValueOnce(Promise.resolve())
     actions.loadAssessments.mockReturnValueOnce(Promise.resolve())
@@ -58,5 +68,15 @@ describe('MainView.vue', () => {
       expect(actions.load).toHaveBeenCalledWith(expect.anything(), 'abcde', undefined)
       done()
     }, 0)
+  })
+  it('updates search term and dispatches filter actions when the search term changes', () => {
+    const wrapper = shallowMount(MainView, { store, localVue })
+
+    wrapper.find(SearchComponent).vm.$emit('seachChanged', 'mini')
+
+    expect(updateSearchTerm).toHaveBeenCalled()
+    expect(filterSections).toHaveBeenCalled()
+    expect(filterSubsections).toHaveBeenCalled()
+    expect(loadGridVariables).toHaveBeenCalled()
   })
 })
